% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrmtbuilder.R
\name{tfrmtbuilder}
\alias{tfrmtbuilder}
\title{Run tfrmt Builder Shiny App}
\usage{
tfrmtbuilder(run = TRUE)
}
\arguments{
\item{run}{Boolean for whether the created object should be run directly. Set to \code{FALSE} for deployment}
}
\value{
Shiny app for creating and modifying tfrmt objects
}
\description{
Run tfrmt Builder Shiny App
}
\examples{
if (interactive()){
  tfrmtbuilder()
}
}
